+++ lib/passage_flutter.dart
@@ -1,19 +1,21 @@
 import 'package:passage_flutter/passage_passkey.dart';
+import 'package:passage_flutter/passage_social.dart';
 import 'passage_flutter_models/auth_result.dart';
 import 'passage_flutter_models/authenticator_attachment.dart';
 import 'passage_flutter_models/passage_app_info.dart';
-import 'passage_flutter_models/passage_social_connection.dart';
 import 'passage_flutter_models/passage_user.dart';
 import 'passage_flutter_models/passkey.dart';
 import 'passage_flutter_platform/passage_flutter_platform_interface.dart';
 
 class PassageFlutter {
   late final PassagePasskey passkey;
+  late final PassageSocial social;
   PassageFlutter([String? appId]) {
     if (appId != null) {
       PassageFlutterPlatform.instance.initWithAppId(appId);
     }
     passkey = PassagePasskey();
+    social = PassageSocial();
   }
 
   // OTP METHODS
@@ -136,18 +138,7 @@ class PassageFlutter {
     return PassageFlutterPlatform.instance.getMagicLinkStatus(magicLinkId);
   }
 
-  // SOCIAL AUTH METHODS
-  Future<void> authorizeWith(PassageSocialConnection connection) {
-    return PassageFlutterPlatform.instance.authorizeWith(connection);
-  }
-
-  Future<AuthResult> finishSocialAuthentication(String code) {
-    return PassageFlutterPlatform.instance.finishSocialAuthentication(code);
-  }
-
-  Future<AuthResult> authorizeIOSWith(PassageSocialConnection connection) {
-    return PassageFlutterPlatform.instance.authorizeIOSWith(connection);
-  }
+ 
 
   // TOKEN METHODS
